+++ lib/puppet/provider/oneview_resource.rb
@@ -27,7 +27,7 @@ module Puppet
       super(value)
       @property_flush ||= {}
       @data ||= {}
-      @client ||= OneviewSDK::Client.new(login)
+      @client ||= client
       @resourcetype ||= ov_resource_type
     end
 
@@ -37,6 +37,10 @@ module Puppet
 
     def oneview_class
       ov_resource_type
+    end
+
+    def client
+      OneviewSDK::Client.new(login)
     end
 
     def self.client
@@ -108,11 +112,11 @@ module Puppet
 
     # Helpers
     def resource_name
-      self.class.to_s.split('::')[2].split('_').drop(1).collect(&:capitalize).join
+      extract_resource_name(self.class.to_s)
     end
 
     def self.resource_name
-      to_s.split('::')[2].split('_').drop(1).collect(&:capitalize).join
+      extract_resource_name(to_s)
     end
 
     def resource_variant